1. Правки по замечаниям
