+++ src/components/film-card.js
@@ -35,7 +35,7 @@ const createFilmCardMarkup = (film) => {
   );
 };
 
-const createFilmCardTemplate = (films, beginCount = 0, endCount) => {
+const createFilmCardTemplate = (films, endCount, beginCount = 0) => {
   const filmMarkup = films.slice(beginCount, endCount)
     .map((film) => createFilmCardMarkup(film)).join(`\n`);
   return `${filmMarkup}`;
