+++ src/mock/const.js
@@ -1,0 +1,57 @@
+export const ACTOR_NAMES = [
+  `Erich von Stroheim`,
+  `Dan Duryea`,
+  `Hugh Grant`,
+  `Jason Statham`,
+  `Mary Beth Hughes`,
+  `Matthew McConaughey`,
+  `Quentin Tarantino`,
+];
+
+export const DESCRIPTION_ITEMS = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra.`,
+  `Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat.  `,
+  `Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat.`,
+  `Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`,
+  `Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus. `,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat.`,
+];
+
+export const DIRECTOR_NAMES = [
+  `Фёдор Бондарчук`,
+  `Anthony Mann`,
+  `Guy Ritchie`,
+  `Peter Farrelly`,
+  `Quentin Tarantino`,
+  `Tom McGrath`,
+];
+
+export const POSTERS_IMAGES = [
+  `./images/posters/made-for-each-other.png`,
+  `./images/posters/popeye-meets-sinbad.png`,
+  `./images/posters/sagebrush-trail.jpg`,
+  `./images/posters/santa-claus-conquers-the-martians.jpg`,
+  `./images/posters/the-dance-of-life.jpg`,
+  `./images/posters/the-great-flamarion.jpg`,
+  `./images/posters/the-man-with-the-golden-arm.jpg`,
+];
+
+export const TITLE_ITEMS = [
+  `The Dance of Life`,
+  `Sagebrush Trail`,
+  `The Man with the Golden Arm`,
+  `Santa Claus Conquers the Martians`,
+  `Popeye the Sailor Meets Sindbad the Sailor`,
+  `The Great Flamarion`,
+  `Made for Each Other`,
+];
+
+export const WRITER_NAMES = [
+  `Anne Wigton`,
+  `Guy Ritchie`,
+  `Heinz Herald`,
+  `Michael McCullers`,
+  `Quentin Tarantino`,
+  `Richard Weil`,
+];
